requirements.txt
google-genai>=1.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
PyGitHub>=2.3.0
gitpython>=3.1.43
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
